Add a FreeBSD feature note to tools
The crt1 provided for FreeBSD binaries contains a feature note section.
It's used to specify OS specific elf properties that are later parsed
in the kernel image activator.
For example one could use it to opt-out of ASLR, or W^X policy.

Since valgrind tools use a custom _start method it has to be added manually.
By default we opt-out from W^X policy, since it's incompatible with
how the profiled programs are loaded.
All features can be modified with elfctl tool that is shipped with
the base system.